+++ angular/lib/src/core/linker/style_encapsulation.dart
@@ -4,8 +4,57 @@ import 'package:angular/src/core/linker/app_view_utils.dart';
 import 'package:angular/src/runtime.dart';
 import 'package:meta/dart2js.dart' as dart2js;
 
+/// Clears all component styles from the DOM.
+///
+/// This should only be called in development mode, typically to reset `<style>`
+/// tags in the DOM between DDC hot restarts or hermetic test cases.
+void debugClearComponentStyles() {
+  if (!isDevMode) {
+    throw StateError(
+      'This function should only be used in development mode.\n'
+      '\n'
+      'See "debugClearComponentStyles()" documentation for details.',
+    );
+  }
+  ComponentStyles._debugClear();
+}
+
 /// Stores `styles: [ ... ]`,  `styleUrls: [ ... ]` for a given `@Component`.
 class ComponentStyles {
+  /// All component `<style>` tags currently in the DOM.
+  ///
+  /// Intentionally stored as [Node] (which is sufficient for this use case) to
+  /// avoid cost of reifying another type (the framework stores lists of [Node]
+  /// elsewhere, but not [StyleElement]).
+  static List<Node> _debugStyleElements;
+
+  /// Callbacks to invoke when [_debugClear] is called.
+  static List<void Function()> _debugClearCallbacks;
+
+  /// See [debugClearComponentStyles].
+  static void _debugClear() {
+    if (_debugClearCallbacks != null) {
+      for (final callback in _debugClearCallbacks) {
+        callback();
+      }
+      _debugClearCallbacks.clear();
+    }
+    if (_debugStyleElements != null) {
+      for (final element in _debugStyleElements) {
+        element.remove();
+      }
+      _debugStyleElements.clear();
+    }
+  }
+
+  /// Registers a [callback] to be called by [_debugClear].
+  ///
+  /// Used to clear static component styles in generated views.
+  static void debugOnClear(void Function() callback) {
+    _debugClearCallbacks ??= [];
+    _debugClearCallbacks.add(callback);
+  }
+
   /// Originating URL of the `@Component`; used in debug builds only.
   final String _componentUrl;
 
@@ -89,7 +138,13 @@ class ComponentStyles {
       target.add('/* From: $_componentUrl*/');
     }
     final styles = _flattenStyles(_styles, target, _componentId).join('\n');
-    document.head.append(StyleElement()..text = styles);
+    final styleElement = StyleElement()..text = styles;
+    // Cache styles in development mode so they can be removed on hot restart.
+    if (isDevMode) {
+      _debugStyleElements ??= [];
+      _debugStyleElements.add(styleElement);
+    }
+    document.head.append(styleElement);
   }
 }